Rename bundle_producer/bundle_processor to system_bundle_producer/system_bundle_processor
For the first implementation of core domain executor, I might copy the
system_bundle_producer/system_bundle_processor as the foundation. We can
refactor them later once we have the mvp of core domain executor.
