+++ domains/client/domain-executor/src/system_bundle_producer.rs
